Rename Drafted to Draft and move under theme/ContentVisibility
