+++ packages/docusaurus-theme-classic/src/theme/BlogTagsPostsPage/index.tsx
@@ -20,7 +20,7 @@ import BlogListPaginator from '@theme/BlogListPaginator';
 import SearchMetadata from '@theme/SearchMetadata';
 import type {Props} from '@theme/BlogTagsPostsPage';
 import BlogPostItems from '@theme/BlogPostItems';
-import Unlisted from '@theme/Unlisted';
+import Unlisted from '@theme/ContentVisibility/Unlisted';
 import Heading from '@theme/Heading';
 
 function BlogTagsPostsPageMetadata({tag}: Props): JSX.Element {
